echoForMe/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
DateTime==4.3
Django==3.1.3
gTTS==2.2.1
idna==2.10
numpy==1.19.4
pandas==1.1.4
playsound==1.2.2
pygame==2.0.0
python-dateutil==2.8.1
python-vlc==3.0.11115
pytz==2020.4
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
webdriver-manager==3.2.2
zope.interface==5.2.0
